+++ ItemQualities/Assets/ItemQualities/Scripts/HealthBarHooks.cs
@@ -1,5 +1,5 @@
-﻿using EntityStates.BrotherMonster;
-using HG;
+﻿using HG;
+using ItemQualities.Utilities;
 using ItemQualities.Utilities.Extensions;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
@@ -41,8 +41,7 @@ namespace ItemQualities
 
             public static readonly FieldInfo[] BarrierOverlayBarInfoFields = new FieldInfo[]
             {
-                typeof(AdditionalBarInfos).GetField(nameof(BarrierBarInfoUnder)),
-                typeof(AdditionalBarInfos).GetField(nameof(BarrierBarInfoOver)),
+                typeof(AdditionalBarInfos).GetField(nameof(BarrierOverflowBarInfo))
             };
 
             public static readonly FieldInfo[] EndBarInfoFields = Array.Empty<FieldInfo>();
@@ -54,12 +53,15 @@ namespace ItemQualities
 
             public readonly HealthBar.BarInfo TemporaryHealthBarInfo;
 
-            public readonly HealthBar.BarInfo BarrierBarInfoUnder;
-            public readonly HealthBar.BarInfo BarrierBarInfoOver;
+            public readonly HealthBar.BarInfo BarrierOverflowBarInfo;
 
             public readonly int EnabledBarCount;
 
-            public AdditionalBarInfos(in HealthBar.BarInfo stealthKitLowHealthUnderBarInfo, in HealthBar.BarInfo stealthKitLowHealthOverBarInfo, in HealthBar.BarInfo temporaryShieldBarInfo, in HealthBar.BarInfo temporaryHealthBarInfo, in HealthBar.BarInfo barrierBarInfoUnder, in HealthBar.BarInfo barrierBarInfoOver)
+            public AdditionalBarInfos(in HealthBar.BarInfo stealthKitLowHealthUnderBarInfo,
+                                      in HealthBar.BarInfo stealthKitLowHealthOverBarInfo,
+                                      in HealthBar.BarInfo temporaryShieldBarInfo,
+                                      in HealthBar.BarInfo temporaryHealthBarInfo,
+                                      in HealthBar.BarInfo barrierOverflowBarInfo)
             {
                 int enabledBarCount = 0;
 
@@ -77,12 +79,46 @@ namespace ItemQualities
 
                 setBarInfo(out TemporaryHealthBarInfo, temporaryHealthBarInfo);
 
-                setBarInfo(out BarrierBarInfoUnder, barrierBarInfoUnder);
-
-                setBarInfo(out BarrierBarInfoOver, barrierBarInfoOver);
+                setBarInfo(out BarrierOverflowBarInfo, barrierOverflowBarInfo);
 
                 EnabledBarCount = enabledBarCount;
             }
+        }
+
+        static readonly Dictionary<Sprite, Sprite> _barToGreyscaleBarrierBarLookup = new Dictionary<Sprite, Sprite>();
+        static readonly HashSet<Sprite> _greyscaleBars = new HashSet<Sprite>();
+
+        static Sprite getGreyscaleBar(Sprite bar)
+        {
+            if (_barToGreyscaleBarrierBarLookup.TryGetValue(bar, out Sprite greyscaleBar))
+                return greyscaleBar;
+
+            if (_greyscaleBars.Contains(bar))
+                return bar;
+
+            Texture2D greyscaleTexture = TextureUtils.CreateAccessibleCopy(bar.texture);
+            greyscaleTexture.name += "_BW";
+
+            Color32[] colors = greyscaleTexture.GetPixels32();
+            for (int i = 0; i < colors.Length; i++)
+            {
+                ref Color32 color = ref colors[i];
+                byte a = color.a;
+                Color.RGBToHSV(color, out float h, out float s, out float v);
+                color = Color.HSVToRGB(h, 0, v);
+                color.a = a;
+            }
+
+            greyscaleTexture.SetPixels32(colors);
+            greyscaleTexture.Apply(true, true);
+
+            greyscaleBar = Sprite.Create(greyscaleTexture, bar.rect, bar.pivot, bar.pixelsPerUnit, 0, SpriteMeshType.FullRect, bar.border);
+            greyscaleBar.name = bar.name + "_BW";
+
+            _barToGreyscaleBarrierBarLookup.Add(bar, greyscaleBar);
+            _greyscaleBars.Add(greyscaleBar);
+
+            return greyscaleBar;
         }
 
         [SystemInitializer]
@@ -91,6 +127,8 @@ namespace ItemQualities
             IL.RoR2.UI.HealthBar.CheckInventory += HealthBar_CheckInventory;
 
             IL.RoR2.UI.HealthBar.ApplyBars += HealthBar_ApplyBars;
+
+            IL.RoR2.HealthComponent.GetHealthBarValues += HealthComponent_GetHealthBarValues;
         }
 
         static void HealthBar_CheckInventory(ILContext il)
@@ -452,28 +490,81 @@ namespace ItemQualities
                 }
             }
 
-            HealthBar.BarInfo barrierBarInfoUnder = addBarrierBar(false);
-            HealthBar.BarInfo barrierBarInfoOver = addBarrierBar(true);
-            HealthBar.BarInfo addBarrierBar(bool over)
+            HealthBar.BarInfo barrierOverflowBarInfo = barrierBarInfoTemplate;
+            barrierOverflowBarInfo.enabled = false;
+
+            if (healthBar.style.barrierBarStyle.enabled)
             {
-                HealthBar.BarInfo overBarrierBarInfo = barrierBarInfoTemplate;
-                overBarrierBarInfo.enabled = false;
-                if (healthComponent.barrier > healthComponent.fullBarrier * (over ? 1 : 2) && healthBar.style.barrierBarStyle.enabled)
+                float normalizedBarrierAmount = healthComponent.barrier / healthComponent.fullBarrier;
+
+                static Color getBarColor(int barIndex)
                 {
-                    int barNum = (int)(healthComponent.barrier / healthComponent.fullBarrier) - (over ? 0 : 1);
-                    overBarrierBarInfo.sprite = ItemQualitiesContent.Sprites.barrierBar;
-                    overBarrierBarInfo.color = QualityCatalog.GetQualityTierDef((QualityTier) ((barNum - 1) % (int) QualityTier.Count)).color;
-                    overBarrierBarInfo.normalizedXMax = Math.Min((healthComponent.barrier / healthComponent.fullBarrier) - barNum, 1f - healthBarValues.curseFraction);
-                    overBarrierBarInfo.enabled = true;
-                    if (!over)
-                    {
-                        healthBar.barInfoCollection.barrierBarInfo.enabled = false;
-                    }
+                    if (barIndex == 0)
+                        return Color.white;
+
+                    return QualityCatalog.GetQualityTierDef((QualityTier)((barIndex - 1) % (int)QualityTier.Count)).color;
                 }
-                return overBarrierBarInfo;
+
+                Sprite greyscaleBarrierBar = getGreyscaleBar(healthBar.style.barrierBarStyle.sprite);
+
+                if (normalizedBarrierAmount > 1f)
+                {
+                    float normalizedOverflowBarSize = normalizedBarrierAmount - (int)normalizedBarrierAmount;
+
+                    barrierOverflowBarInfo.enabled = true;
+                    barrierOverflowBarInfo.sprite = greyscaleBarrierBar;
+                    barrierOverflowBarInfo.color = getBarColor((int)normalizedBarrierAmount);
+                    barrierOverflowBarInfo.normalizedXMin = 0f;
+                    barrierOverflowBarInfo.normalizedXMax = normalizedOverflowBarSize * (1f - healthBarValues.curseFraction);
+                }
+
+                // We are displaying more than 2 full bars, change color of normal barrier bar to match previous color
+                if (normalizedBarrierAmount > 2f)
+                {
+                    healthBar.barInfoCollection.barrierBarInfo.sprite = greyscaleBarrierBar;
+                    healthBar.barInfoCollection.barrierBarInfo.color = getBarColor((int)normalizedBarrierAmount - 1);
+                }
+                else if (healthBar.barInfoCollection.barrierBarInfo.sprite == greyscaleBarrierBar)
+                {
+                    // Reset base barrier bar if we don't need custom color
+                    healthBar.barInfoCollection.barrierBarInfo.sprite = healthBar.style.barrierBarStyle.sprite;
+                    healthBar.barInfoCollection.barrierBarInfo.color = Color.white;
+                }
             }
 
-            return new AdditionalBarInfos(stealthKitLowHealthUnderBarInfo, stealthKitLowHealthOverBarInfo, temporaryShieldBarInfo, temporaryHealthBarInfo, barrierBarInfoUnder, barrierBarInfoOver);
+            return new AdditionalBarInfos(stealthKitLowHealthUnderBarInfo, stealthKitLowHealthOverBarInfo, temporaryShieldBarInfo, temporaryHealthBarInfo, barrierOverflowBarInfo);
+        }
+
+        static void HealthComponent_GetHealthBarValues(ILContext il)
+        {
+            ILCursor c = new ILCursor(il);
+
+            int patchCount = 0;
+
+            while (c.TryGotoNext(MoveType.After,
+                                 x => x.MatchLdarg(0),
+                                 x => x.MatchLdfld<HealthComponent>(nameof(HealthComponent.barrier))))
+            {
+                c.Emit(OpCodes.Ldarg_0);
+                c.EmitDelegate<Func<float, HealthComponent, float>>(clampBarrier);
+
+                [MethodImpl(MethodImplOptions.AggressiveInlining)]
+                static float clampBarrier(float barrier, HealthComponent healthComponent)
+                {
+                    return Math.Min(barrier, healthComponent.fullBarrier);
+                }
+
+                patchCount++;
+            }
+
+            if (patchCount == 0)
+            {
+                Log.Warning("Failed to find patch location");
+            }
+            else
+            {
+                Log.Debug($"Found {patchCount} patch location(s)");
+            }
         }
     }
 }